🦄 refactor(context): change context file name
